Factor out hook to filter traces
